Tests/LibWeb/Text/expected/Editing/execCommand-selectAll.txt

No range.
DIV 0 - DIV 0
BODY 0 - BODY 5
